secure delete bookings/id route

+++ routes/bookingsRoutes.js
@@ -61,12 +61,22 @@ router.get('/bookings/:booking_id', async (req, res) => {
 
 router.delete('/bookings/:booking_id', async (req, res) => {
   try {
-    let { rows } = await db.query(
-      `DELETE FROM bookings WHERE booking_id = ${req.params.booking_id} RETURNING *`
+    const decodedToken = jwt.verify(req.cookies.jwt, jwtSecret)
+    if (!decodedToken) {
+      throw new Error('Invalid authentication token!')
+    }
+    let { rowsSelect } = await db.query(
+      `SELECT * FROM bookings WHERE booking_id = ${req.params.booking_id}`
     )
     if (!rows.length) {
       throw new Error(`No booking found with id ${req.params.booking_id}`)
     }
+    if (rowsSelect[0].user_id !== decodedToken.user_id) {
+      throw new Error('You are not authorized to delete this booking!')
+    }
+    let { rows } = await db.query(
+      `DELETE FROM bookings WHERE booking_id = ${req.params.booking_id} RETURNING *`
+    )
     res.json(rows[0])
   } catch (err) {
     res.json({ error: err.message })
